update nhap tien o cac chuc nang bi crash tren dien thoai

+++ app/src/main/java/com/example/doantotnghiep/adapter/AdapterKeHoachTietKiem.java
@@ -55,7 +55,7 @@ public class AdapterKeHoachTietKiem extends ArrayAdapter<ArrayKeHoachTietKiem> {
         return view;
     }
 
-    public static String DoiSoSangTien(Double so) {
+    public static String DoiSoSangTien(int so) {
         DecimalFormat decimalFormat = (DecimalFormat) NumberFormat.getInstance(Locale.US);
         decimalFormat.applyPattern("#,###,###,###");
         return decimalFormat.format((so)) + " đ";
