+++ app/src/main/java/com/example/doantotnghiep/adapter/AdapterVi.java
@@ -47,7 +47,7 @@ public class AdapterVi extends ArrayAdapter<ArrayVi> {
         return view;
     }
 
-    public static String DoiSoSangTien(Double so) {
+    public static String DoiSoSangTien(int so) {
         DecimalFormat decimalFormat = (DecimalFormat) NumberFormat.getInstance(Locale.US);
         decimalFormat.applyPattern("#,###,###,###");
         return decimalFormat.format((so)) + " đ";
