+++ app/src/main/java/com/example/doantotnghiep/model/ArrayLichSuChuyenTien.java
@@ -3,9 +3,9 @@ package com.example.doantotnghiep.model;
 public class ArrayLichSuChuyenTien {
     public String thoigian, vichuyen, vinhan;
     public int ma;
-    public Double tien;
+    public int tien;
 
-    public ArrayLichSuChuyenTien(String thoigian, String vichuyen, String vinhan, int ma, Double tien) {
+    public ArrayLichSuChuyenTien(String thoigian, String vichuyen, String vinhan, int ma, int tien) {
         this.thoigian = thoigian;
         this.vichuyen = vichuyen;
         this.vinhan = vinhan;
