+++ app/src/main/java/com/example/doantotnghiep/model/ArrayThuChiChoKeHoachTietKiem.java
@@ -2,10 +2,10 @@ package com.example.doantotnghiep.model;
 
 public class ArrayThuChiChoKeHoachTietKiem {
     public int mathuchichokehoach;
-    public Double sotienthuchichokehoach;
+    public int sotienthuchichokehoach;
     public String loaithuchichokehoach, motathuchichokehoach, ngaythuchienthuchichokehoach;
 
-    public ArrayThuChiChoKeHoachTietKiem(int mathuchichokehoach, Double sotienthuchichokehoach, String loaithuchichokehoach, String motathuchichokehoach, String ngaythuchienthuchichokehoach) {
+    public ArrayThuChiChoKeHoachTietKiem(int mathuchichokehoach, int sotienthuchichokehoach, String loaithuchichokehoach, String motathuchichokehoach, String ngaythuchienthuchichokehoach) {
         this.mathuchichokehoach = mathuchichokehoach;
         this.sotienthuchichokehoach = sotienthuchichokehoach;
         this.loaithuchichokehoach = loaithuchichokehoach;
